+++ oscar_docdata/__init__.py
@@ -1,3 +1,3 @@
 # following PEP 386
 # Will keep "1.2" prefix to be in sync the API version.
-__version__ = "1.2.4"
+__version__ = "1.2.5a1"